requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-engineio==4.9.0
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
propcache==0.3.2
PyJWT==2.10.1
python-dotenv==1.1.1
redis==6.3.0
requests==2.32.4
SQLAlchemy==2.0.41
twilio==9.7.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
